Pause and reset channels when load driver
It's an update to the channel initialization sequence.

A reset to the channels is necessary during driver initialization. However, it
could be dangerous to reset a channel with outstanding SE processing by the DMAC
in the background. To avoid this, we pause the channel and wait 60ms to ensure
that all the SE processing has been completed before we reset it during driver
initialization.