Remove static modifier for the ce variable in function process_desc
It was observed that when running multiple channels, sometimes the
CE cannot be completed properly (dma_sync_wait failed). It turned out to
be the variable ce in function switchtec_dma_process_desc was declared
static which makes the CE processing in one channel affect the others.
Fix it by removing static.